move IUser in security package
