+++ back/src/main/java/org/crunchytorch/coddy/user/elasticsearch/entity/UserEntity.java
@@ -1,6 +1,6 @@
 package org.crunchytorch.coddy.user.elasticsearch.entity;
 
-import org.crunchytorch.coddy.user.data.IUser;
+import org.crunchytorch.coddy.security.data.IUser;
 import org.crunchytorch.coddy.user.data.in.UpdateUser;
 import org.crunchytorch.coddy.security.utils.SecurityUtils;
 import org.springframework.data.annotation.Id;
